refactor: confirm modal in shared components
