+++ decky-plugin/src/shared/components/confirm.tsx
